Move on_command and on_command_completion timing to heartbeat.py
